requirements.txt
Flask
Flask-Cors
cryptography
numpy
Pillow
soundfile
moviepy
opencv-python-headless